requirements.txt
graphviz==0.21
PyYAML==6.0.2

pydantic==2.7.1
mkdocs==1.6.0
mkdocs-material==9.5.19

mkdocs-git-revision-date-localized-plugin
mkdocs-glightbox
mkdocs-minify-plugin
mkdocs-redirects
